Move hardcoded strings to separate class.
