--- src/main/java/com/altinity/clickhouse/sink/connector/validators/ConfigValidator.java
@@ -1,5 +1,0 @@
-package com.altinity.clickhouse.sink.connector.validators;
-
-public class ConfigValidator {
-
-}
